+++ src/bika/lims/permissions.py
@@ -113,6 +113,7 @@ TransitionSampleSample = "senaite.core: Transition: Sample Sample"
 TransitionScheduleSampling = "senaite.core: Transition: Schedule Sampling"
 TransitionDispatchSample = "senaite.core: Transition: Dispatch Sample"
 TransitionRestoreSample = "senaite.core: Transition: Restore Sample"
+TransitionCreatePartitions = "senaite.core: Transition: Create Partitions"
 
 
 # Type-specific permissions