+++ src/senaite/core/upgrade/v02_03_000.py
@@ -21,7 +21,9 @@
 from bika.lims import api
 from Products.Archetypes.config import REFERENCE_CATALOG
 from senaite.core import logger
+from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.core.config import PROJECTNAME as product
+from senaite.core.setuphandlers import _run_import_step
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
 
@@ -44,7 +46,16 @@ def upgrade(tool):
     logger.info("Upgrading {0}: {1} -> {2}".format(product, ver_from, version))
 
     # -------- ADD YOUR STUFF BELOW --------
+
+    # run import steps located in bika.lims profiles
+    _run_import_step(portal, "rolemap", profile="profile-bika.lims:default")
+
+    # run import steps located in senaite.core profiles
+    setup.runImportStepFromProfile(profile, "rolemap")
+    setup.runImportStepFromProfile(profile, "workflow")
+
     fix_worksheets_analyses(portal)
+    fix_cannot_create_partitions(portal)
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
@@ -96,3 +107,25 @@ def is_orphan(uid):
     """
     obj = api.get_object_by_uid(uid, None)
     return obj is None
+
+
+def fix_cannot_create_partitions(portal):
+    """Updates the role mappings of samples in status that are affected by the
+    issue: sample_received, verified, to_be_verified and published statuses
+    """
+    logger.info("Fix cannot create partitions ...")
+    wf_tool = api.get_tool("portal_workflow")
+    workflow = wf_tool.getWorkflowById("senaite_sample_workflow")
+    statuses = ["sample_received", "to_be_verified", "verified", "published"]
+    query = {
+        "portal_type": "AnalysisRequest",
+        "review_state": statuses,
+    }
+    brains = api.search(query, SAMPLE_CATALOG)
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        if num and num % 10 == 0:
+            logger.info("Fix cannot create partitions {0}/{1}".format(num, total))
+        obj = api.get_object(brain)
+        workflow.updateRoleMappingsFor(obj)
+        obj.reindexObject(idxs=["allowedRolesAndUsers"])